third_party/WebKit/LayoutTests/http/tests/local/fileapi/send-dragged-file-form-iso-2022-jp.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<meta charset="utf-8">
<title>Send a dragged UTF-8-named file via POST to an ISO-2022-JP form</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#multipart-form-data">
<link rel="author" title="Benjamin C. Wiley Sittler"
      href="mailto:bsittler@chromium.org">
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async testCase => {

  if (document.readyState !== 'complete') {
    await new Promise(resolve => addEventListener('load', resolve));
  }
  assert_own_property(
      window,
      'eventSender',
      'This test requires the test framework to expose eventSender for File drag-and-drop event injection');

  const fileToDrop =
        '../resources/file-for-drag-to-send3-ABC~¤•★星🌟星★•¤~XYZ.txt';
  const baseName = fileToDrop.split('/').pop();
  const encoding = 'iso-2022-jp';
  const expectedEncodedBaseName =
        'file-for-drag-to-send3-ABC~&#164;&#8226;\x1b$B!z@1\x1b(B&#127775;\x1b$B@1!z\x1b(B&#8226;&#164;~XYZ';

  let targetLoaded, resolveTargetLoaded;
  targetLoaded = new Promise(resolve => {
    resolveTargetLoaded = resolve;
  });
  const target = Object.assign(document.createElement('iframe'), {
    name: 'target',
  });
  document.body.insertBefore(target, document.body.firstChild);

  const form = Object.assign(document.createElement('form'), {
    acceptCharset: encoding,
    action: 'http://127.0.0.1:8000/xmlhttprequest/resources/post-echo.cgi',
    method: 'POST',
    enctype: 'multipart/form-data',
    target: target.name,
  });

  const fileInput = Object.assign(document.createElement('input'), {
    type: 'file',
    name: 'file',
    ondragenter: () => event.preventDefault(),
    ondragover: () => event.preventDefault(),
    ondrop: () => {
      assert_not_equals(
          event.dataTransfer.types.indexOf('Files'),
          -1,
          'event.dataTransfer must contain a File object on drop.');
      event.preventDefault();
      fileInput.files = event.dataTransfer.files;
      target.onload = () => resolveTargetLoaded();
      form.submit();
    },
  });
  fileInput.style.cssText = 'width: 100px; height: 100px';

  form.appendChild(fileInput);

  // Important that we put this element at the top of the doc so that logging
  // does not cause it to go out of view (where it can't be dragged
  // to)
  document.body.insertBefore(form, document.body.firstChild);

  // TODO: switch to chrome.gpuBenchmarking.pointerActionSequence once
  // it supports file dragging; even better would be a FileList
  // constructed using a contents Blob and a filename string, but that
  // is also not yet possible.
  eventSender.beginDragWithFiles([fileToDrop]);
  const centerX = fileInput.offsetLeft + fileInput.offsetWidth / 2;
  const centerY = fileInput.offsetTop + fileInput.offsetHeight / 2;
  eventSender.mouseMoveTo(centerX, centerY);
  eventSender.mouseUp();

  await targetLoaded;
  const responseText = target.contentDocument.body.innerText;
  const responseLines = responseText.split('\n');
  if (responseLines.length && !responseLines[responseLines.length - 1]) {
    --responseLines.length;
  }
  assert_greater_than(
      responseLines.length,
      2,
      'Response not multipart-shaped (too few lines):\n' + JSON.stringify(responseText));
  assert_equals(
      responseLines[responseLines.length - 1],
      responseLines[0] + '--',
      'Response not multipart-shaped (missing expected boundaries):\n' + responseText);
  const boundary = responseLines[0];
  const expectedText = [
    boundary,
    'Content-Disposition: form-data; name="file"; filename="' +
        expectedEncodedBaseName +
        '"',
    'Content-Type: text/plain',
    '',
    'ABC~¤•★星🌟星★•¤~XYZ',
    '',
    boundary + '--',
  ].join('\n');
  assert_equals(
      responseText.indexOf(expectedText),
      0,
      'Unexpected multipart-shaped response data received:\n' +
        responseText +
        '\nExpected:\n' +
        expectedText);
}, 'Send a dragged UTF-8-named file via POST to an ISO-2022-JP form');

</script>
